Allow flow definitions to be referenced in production build
This is a working example and includes code that shouldn't be merged
into electron-react-boilerplate.

Note that by default, flow-runtime only runs assertions when
`process.env.NODE_ENV === 'development'`. This change makes Flow types
accessible to code, but assertions just aren't run. This is a
configuration that is recommended for production in the flow-runtime
docs.